api/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
python-decouple==3.3
psycopg2==2.8.6
